Update src contents
