+++ src/index.html
@@ -12,8 +12,8 @@
         <header>
             <h3>Today's &nbsp; To Do<i class="fas fa-sync-alt"></i></h3>
         </header>
-        <form class="add-task">
-            <input id="form-id" type="text" placeholder="Add your task here ...">
+        <form id="form-id" class="add-task">
+            <input  type="text" placeholder="Add your task here ...">
             <i class="fas fa-level-down-alt fa-rotate-90 icon"></i>
             <input class="submit" type="submit" hidden>
         </form>
