+++ src/style.css
@@ -96,6 +96,40 @@ label {
 .list-item {
   display: flex;
   align-items: center;
-  gap: 10px;
+  gap: 1rem;
   padding-left: 0.5rem;
+}
+
+.list-input {
+  transform: scale(1.3);
+}
+
+.item-details {
+  height: 2rem;
+  padding: 0.5rem;
+  width: 300px;
+  border: none;
+  resize: none;
+  outline: none;
+}
+
+.fa-trash-alt {
+  display: none;
+  margin-right: 2rem;
+  color: rgba(83, 83, 83, 0.7);
+}
+
+.fa-square {
+  font-size: 1.1rem;
+  margin-top: 1.1rem;
+  color: rgb(83, 83, 83);
+}
+
+.clear {
+  text-align: center;
+  font-size: 18px;
+  color: rgb(104, 96, 96);
+  background-color: #91909048;
+  margin-bottom: 0;
+  padding: 20px 0;
 }
